Use Builtin class summary_text in `?`
Recenty we have been filling out `summary_text` for Built-in Functions.

This text currently appears in Django when it is listed in a group.
See for example "Operations on Strings".

This PR expands that use so that this short text also appears
when `?xxx` or Information[xxx, LongForm->False]` is requested.