Revove duplicated function The function `_clusters` was declared twice, with the exactly same body
Remove duplicated function: create_css

Remove useless sort
The tick numbers were being sorted, but as they are independent, it
isn't necessary

Remove useless argument
In the method `extent` from the class `Style`, there were an argument
`pre`, which by default was `True`. It was never set to `False`.
Anyway, the order of the style doesn't matter

Remove useless None checking
Stop checking if the color is None, this already was started before,
but this place was forgotten

Stop converting color into list