Reduce unnecessary expansion & untangle imports...
A number of different kinds of atoms that can't change value
unnecessarily go through an extra expansion for fixpoint check.

In particular many constants values: strings, numbers, fixed symbols
e.g. True, False, N, etc.

The one place where we *do* need to check though is Symbol, becuase a
value like "x" may change. So we separate Symbols from Constants.

Also we remove "import" of a name where that name isn't used directly in
the module. This seems to happen so that some other module can find the
name instead of where it was defined. This isn't helpful. In fact, it is
the opposite.