Revise Expression to use ElementsProperties
ElementsProperties is a mutable record.

Note that addition work is needed to categorize existing Expression() calls into `to_expression()` or
`Expression()`.

After that is done, we should see a big improvement in speed due to a
reduction of unneeded calls to convert elements.

Also, there should be improvements as we start setting the elements
properties in the calls to be more precise.

Until then, the performance here will be slightly worse with this
commit, since we have the worst of both worlds: we have to do conversion
on each Expression() call as well as conversion that may already be done in setup to calling
Expression which will be required (not optional) in the future.

A specialization of Expression, ListExpression() will be using this as the base code to work off of.