Remove a weirdness
One of the things I find in looking over the code is that since
historically programmers have been loose in Expression arguments we get
stuff like this.

In other places, I have seen we have conversions from one format to
another only to have it unconverted later in another conversion.