Speed up tests by reducing numbers...
Here there are a number of tests where having large numbers really is
showning much benefit. For example whether we wan the 500 digit versus
the 1000000 digit is kind of pointless and just takes time.

Or how about a maximum number of iterations. In the end we get this same
result: too many iterations.

For plotting and image processing though, I suppose there is benefit in
taking the time since in the end you see a prettier plot. In the case of
images, well, we use reasonable sinze images.

(The time for plotting eventually I expect to see drop.)