Reverse Cython default.
Note that for Pyston this slow things down.
For PyPy we can't use it either.

There is benefit in doing this over CPython (probably up to 3.9), but
since it takes minutes to run Cython, for development this often isn't a
big win since it can happen that files you are working on change and so you'll need to
run again, with the 2 minutes again. Or files will stale, sometimes
inadvertently.

So running Cython is worth the effort before building a
wheel (i.e. before release) or using in a docker container.