Correct  order of parameters in asy rectanglebox ..
this caused the Dice example in the manual to fail - it uses
rectanglebox.

Also, add type annotations on functions.

I think when one of the opacity parameters was added not all of the
routines were adjusted. And we didn't have a test to check this either.
We have a test now.