Rename lower-level tex formatters to latex...
since that is what it is, and we are striving for better clarity.

I am still being a little loose here since LaTeX extensions are ".tex"
instead of ".ltx" or ".latex". And I am not insisting on LaTeX instead of
latex. (But, who knows that may happen someday).

For compatibility with outside projects, e.g. Django, a symbolic link
from mathics/doc/tex is to mathics/doc/latex has been added.