Start to make use of literalness...
to reduce conversions

If an object is_literal, then its value is the Python representation.

Although there may be benefit in noting literals that do not have direct
Python representations like Complex (as it currently is implemented) or
Boxes, the vast majority of literals that we can make immediate use of
is where there is a direct Python representation, e.g. Integers, Reals,
Strings, and ListExpressions (of literals).

Also, go over some types and remove a few more get_string_value()'s.