Move mathics.core.builtin code into system_init...
Although we now require a frontend to *explicitly* call
``initialize_system()`` and this may seem more akward, it is helpful
and important for a couple of reasons.

First, it provides better modularization and reduces circular imports:
``mathics.builtin``  is a close to top-level import and, in contrast to
``mathics.builtin.system_init`` there are a lot of modules underneath
``mathics.builtin``. (There are none under
``mathics.builtin.system_init``).

As a result we can have more ``import mathics.builtin.system_init`` from a module
where we can't do the same for ``import mathics.builtin``.

Second, by placing initialization of structures which do not
change outside of development we are in a better position to have write
and admninstrative tool to load all of this and dump it to as a system
image, which can be read in rather than built it up from scratch as
happens now. This new function ``initialize_system`` might have a parameter
to read the image file rather than do the work it does now.

In its current form though ``initialize_system()`` does as not
encapsulate system loading as much as it could.