storing attributes when a pattern is created (#804)
Inspired in discussion #800, this changes a little bit the
implementation of `Pattern`, in a way that attributes are stored in the
object instead of looking at it in the Definitions object. Just with
this, doctest cuts other 7s (~7% faster than master in Pyston).
Also, I pytests with the tests proposed in #800 were included.