adding Elements. Local symbols in rules/Bessel.m  inside a private context. (#818)
This PR adds the built-in `Element`, which is used in defining certain
integrals. In particular, it would be useful to extend more the support
for integral representations of special functions.

Also, I noticed that when `Set`/`SetDelayed` expressions with a LHS
involving named patterns appear in autoload files, the temporal names
can spoil the `System` namespace, making some test fail. A way to
prevent that is to enclose the code of the modules in a `Begin`/`End`
statement. Public symbols defined in the module appear then before the
private block, by assigning a usage string.