Moving private doctests to pytests. Fix the compatibility issue 0. I-> 0. (#895)
In this PR, I start to move some private doctests to Pytest. I also
propose some extra related tests and check that the expected behavior
matches with obtained in WMA.
From this, it came up that we are handling wrong how ``` 0. I``` is
evaluated: In master, it is evaluated to `0.`, but in WMA, it is
evaluated to ``` Complex[0.`, 0.`] ```.
To match this behavior, I had to modify `from_mpmath`, and deactivate
its lru_cache. The problem now with lru_cache is that `mpmath.mpf(0.)`
produces the same hash than `mpmath.mpc(0., 0.)`, which looks wrong.