handle box errors in graphics and graphics3d (#966)
Here we go with a proposal to address #964. The idea is that when an
invalid graphics primitive is processed, a pink rectangle is drawn
instead. Notice that in WMA no error message is shown in the client.
Just the pink background is shown in the image, plus a tooltip in the
notebook interface.