Docpipeline tests by chapter and section (#1022)
This continues #1021, and starts to tackle the problem of accessing
tests by chapter and section without interacting over the full
documentation.

For chapters, the change is quite straightforward. On the other hand,
for sections, I think we need to think a little bit what is the best way
to do this.

One possibility would be to produce the documentation entry from the
docstring of the target builtin, without passing through the
documentation. The other way to go would be to keep a dictionary in the
documentation, storing all the sections "by slug".

Thoughts?

---------

Co-authored-by: rocky <rb@dustyfeet.com>