sympy  1.13 compatibility (#1038)
And here we have a partial compatibility patch. The root of the issues
seems to be the change in which comparisons between `sympy.Float` and
`float` are done from sympy 1.13.
It breaks several things, including the cache mechanism for complex
numbers. This draft makes it works by disabling the cache.