Convert RootSum to SymPy (#1136)
I split this out into a separate PR as I think it's going to require
some more work. We already had a SymPy -> Mathics conversion for
RootSum, this adds one going the other way too. However, it results in
some tests failing as it means that Simplify automatically expands
RootSums now, not sure if we want to add some hints to prevent that from
happening.