Workaround a mathics_scanner fragility...
We have a dependency problem in mathics-scanner where
code is performed in tokeniser loading. And the
script that creates a needed JSON file imports
mathics_scanner which imports the tokenser which
needs that JSON file to pre-exist.

The workaround here is to not import is_symbol_name from
mathics_scanner but from mathics_scanner.tokeniser

By doing this when the JSON-creating program imports mathics_scanner
it does not import the module that requires the JSON file.

An attempt was made to delay JSON-reading, but something is still
whacky, so that doesn't work yet. Punt on this aspect for later.