Bugfix in `With`: Return evaluated result (#1194)
# Problem
```mathematica
Block[{i = 0}, With[{}, Module[{j = i}, Set[i, i+1]; j]]]
```` 
returns `1` whereas the correct answer is `0`. This is because, `With`
evaluates the `Module` expression but erroneously returns the
unevaluated version, triggering a second evaluation.

# Fix
The fix is to return the evaluated version. With the fix, the above code
returns `0` as expected.