Adjust sizes and text to match Mathematica better? (#1231)
graphics.py: Reduce size on PointSize, Thickness, AbsoluteThickness
svg.py: reduced thickness of text in InsetBox

I have been looking at graphs in Implementing Discrete Mathematics and
noticed for a while that dots shown in Mathics3 seem a bit too big
compared to what they are in the text, and text for axes is too bold.

Here is an attempt to make this look more like what is printed in this
book.

For comparison with change: examples of dot sizes: 

Before:


![image](https://github.com/user-attachments/assets/5a2ede1b-3220-4877-a509-8b177c3265c8)

Proposed change:


![image](https://github.com/user-attachments/assets/3448f59d-1b0b-40c2-a126-a4da5520c15a)

And text in axes of graphs. Before: 


![image](https://github.com/user-attachments/assets/41636e1f-c8c2-462e-8a71-8bea4b96b7aa)



Proposed change: 


![image](https://github.com/user-attachments/assets/632d00ab-1066-4d71-973d-2bf892f8cede)

Since I don't have access to anything that draws corresponding
Mathematica graphs I rely on you to let me know if this is closer.

Note this is not about whether one is aesthetically better than the
other, but instead, whether this matches what Mathematica does. The
examples are all taken from the Django gallery.