Combinatorica as submodule (#1226)
Remove Combinatorica from this repository but include it as a git
submodule from Mathics3-Combinatorical.

As a by-product, version 2.0.0 of Combinatorica is also now available.
An additional CI job is set up to run the Combinatorica tests that we
have (currently for v0.9). These will be expanded soon.
 And possibly we may include v2.0.0 tests as well. 

To be determined if we want the one CI just to run tests for other
packages as well.