Start adding Alignment to InsetBox in 2D graphs (#1234)
Add the notion of text alignment of InsetBox which is typically used in drawing axes. The text for an x (horizontal) axis is under the ticks while the text for a y (vertical) axis is to the left.

Here we've done this for asymptote and SVG. More that could be done, but this is a small start.