Parsing refactor (#1260)
Originally I wanted to make progress on Box operator parsing. However as happens too often, before I can make forward progress it is useful to clean things up and document the current state.

This is not getting a bit large, so probably we should do the clean up work separate from the box operators to follow.