Reinforce in get_functions that eval/format methods are indeed callable (#1322)
`Builtin.get_functions` collects attributes of the subclasses with names starting with a prefix and it assumes that these are methods. 

Until now, the way to check that they are indeed methods was to check if they have a docstring. What we want to check is whether the method is callable. In Python 3.13, this is needed.
This PR implements a callable check.