More robust docstring parsing (#1323)
In the current (master) documentation/doctest implementation, it is assumed that all the docstrings have a "margin" of spaces, from the class indentation. Then, the regular expression looking for doctests
fails to find them if there is not at least a single space at the beginning of the line.

Python 3.13 removes the "left margin" of docstrings, so the regular expression will fail to detect some tests.
This PR changes the regular expression to allow doctests which do not start with a space.

Also, in future versions of Python, tab characters in docstrings will be converted into spaces, which will make to fail some tests. In this PR, tabs in doctests are replaced by sequences of 4 spaces, making the tests more robust under this kind of change.