Revise API for position tracking (#1413)
Start tracking locations in specific expressions to aid error reporting
and debugging.

We need to take care to avoid bloating memory with positions.

Right now, start locations are tagged for FunctionApplyRule, an ExpressionPattern -> Python method. Here, the location is the Python method, or rather, can be extracted from that. So we can reuse the location information that Python is storing anyway.

For Expressions and ExpressionPatterns that are read in from Files, Things are more complicated..

This needs to be coordinated with a change in mathics-scanner with the same git branch name.