Introduce get_eval_Expression()
get_eval_Expression can be used in Mathics3 Builtin evaluation methods to find the Expression that caused the evaluation get triggered.

It is often needed in error messages of "eval()" methods to report what Expression is in error.

It is fragile because it relies on the Mathics3 implementation having a Expression.rewrite_apply_eval_step() method.