Processing options in patterts with Optional values and OptionsPattern together. (#1433)
Digging on the issues with `Set*` I found some issues on how we store
and apply pattern rules during evaluation. Here there is a bunch of
cases of definitions that in the master branch passes the tests just
because misimplementations in the way the rules are stored, sorted and
applied.
There are more changes to come, but these are easy. Most of them comes
from the way in which `Optional` and `OptionsPattern` interact. The new
rules are compatible with the WMA evaluation loop.