fix 1432 (#1469)
This PR fixes the issue #1432 by requiring the escape characters in front of dollar characters in URLs. The escape is necessary at the level of docstrings to avoid the confusion with the open character for a LaTeX equation. Once the expression is parsed, and we are inside a `<URL>` tag, it is safe to remove the escape character.