Shorten TraceEvaluation
Lines of the form:

Replacing: x -> x

don't add anything, so remove them.

In the future, maybe we can have a "verbosity level" to include
something like this. The only purpose I can see, is to indicate that a
rewrite rule was considered and did nothing. However, TraceEvaluations
are typically very long and in my experience doesn't add clarity, just clutter.