Simplfy invalid-argumnent-count pytest testing
There was way too much verbiage and clutter around running pytests to
look check that we are giving a message calling Builtin Function with
the wrong number of arguments.

Also, in pytest -s output, show the failure reason. This makes it
easier to find an error in the pytest file.