Revise Complex represetntation.
Complex.value is now something that appears to be complex in Python.
Complex.exact_value, however is an internal exact representation. It
contains precision values and can be a Mathics3 Integer or Real,
MachineReal number.

Property is_exact_value indicates whether a value is exact when
converted to a Python complex number or not.

Some properties and tests are now cached.