Adjust First[] handling for ByteArray...
For now, we will not try to generalize how Atoms with "elements" or
whatever they might eventually be called is handled. Instead we will
limit to ByteArray which is what we know about now.

This test might get adjusted in the future. But that is then, and this
is now.