Using upvalues of atoms in evaluations. (#1507)
In the `Expression.rewrite_apply_eval_step` method, in the upvalues
lookup, to reduce the cost of evaluating expressions with many numerical
elements, we use to skip atomic elements. This makes that upvalues of
atomic elements get never evaluated (see #1506). This PR fix by, instead
of skipping atoms, just check beforehand if all the elements are of the
same kind.

---------

Co-authored-by: R. Bernstein <rocky@users.noreply.github.com>