Refactor plot3d (#1520)
This PR substantially refactors Plot3D and DensityPlot (which share a
good bit of code), in preparation for adding a new vectorized
implementation of those two functions.  No function is changed or
added.

* The the flow from builtin to eval is cleaner - no more callbacks
  from eval to builtin. Options processing is done prior to calling
  eval_*. Plot3D and DensityPlot now have their own eval functions
  which share implementation code in eval rather than in builtin.

* A PlotOptions class has been introduced. This simplifies passing
  plotting options from builtin to eval. It may also be reusable for
  other plotting builtins.

* A new GraphicsGenerator class has been added. This consolidates and
  regularizes some scattered code for generating Graphics and
  Graphics3D expressions, which are rather complex in structure. This
  will be further enhanced to support GraphicsComplex in the PR that
introduces vectorized plotting functions. Possibly could be retrofitted
  to other plotting functions.

* A test for DensityPlot was added, which was missed in the previous
  PR because I didn't realize how much code it shared with Plot3D.

This is a fairly big change, but I worked carefully, frequently
testing using the detailed tests for Plot3D and DensityPlot.  The
tests caught numerous errors as I worked, increasing confidence that
the warning light was hooked up and no functional changes have been
introduced.

Some quick timing tests show no change in performance as far as I can
see.

More can be done, particularly in the areas of error checking and
typing, but that shouldn't change the PR in any major way so hoping
to get some eyes on it in parallel.