New plot compile functions 2 (#1530)
* Added an optional keyword argument `raise_on_error` to `to_sympy` and
`expression_to_sympy` that causes it to raise the specified exception in
cases where it would otherwise silently have returned `None`. This is
used by `plot_compile` to aid in identifying the cause of compilation
failure, which will help to build out compilation functionality. That is
the only place it is specified, and if it is not specified, behavior is
as it was before.
* Added compilation test cases in `test_plot_compile.py` for `If`,
`Switch`, and `Which`. These are defined in `procedural.py`, but they
are not really procedural.
* Enabled `Mod`, `FractionalPart`, `IntegerPart`, `If`, `Which`, `Min`,
and `Max` for compilation by sympy-enabling them: subclass from
`SympyFunction` and implement `to_sympy`. Verified that this enables
compilation using the test cases in `test_plot_compile.py`, and by doing
some plots:

<img width="726" height="983" alt="test-030"
src="https://github.com/user-attachments/assets/05bd5aec-bf12-41bb-a021-d4b5c84491cf"
/>

---------

Co-authored-by: Juan Mauricio Matera <matera@fisica.unlp.edu.ar>