Better plot tests (#1541)
We only have a couple plot tests that actually look at the output, and
those were constructed and run laboriously, and none of them test the
new vectorized plot code. This makes refactoring the code risky (and so
far I've been relying mostly on end-to-end testing that compares the
generated visuals in my test front-end).

Here's a different approach: the idea is to write the result expression
to a file in outline tree form, and then compare the actual result with
an expected reference result using diff. This provides easily
constructed tests with a fairly readable indication of what has changed.
See comment in test_plot__detailed.py

In addition, this change
* Makes it possible to change at runtime whether vectorized or classic
plot code is used
* Augments print_expression_tree to print the abbreviated numpy arrays
(default numpy str implementation) so that numeric data can be checked
efficiently for gross changes.