Add plot doctests (#1542)
My understanding is that the doctests for plotting don't check the
generated output and only look for the "-Graphics-" string, is that
correct?

Here's a proposal for tests that compare the generated Graphics* output
with expected output. The tests in `test_plot_detail.py` are now
augmented with about 125 plot-related doctests, with corresponding
expected reference output in `test_plot_detail_ref/doc-*`.

I hope this will give us a lot more confidence to refactor plotting code
without breaking anything in the doc. These tests take about 10 seconds
to run so having them separated out from the doctests will make
development cycles faster.

Couple notes:
* I chose to store the test definitions in a yaml file `doc_tests.yaml`
because with that number of tests a Python data structure is a bit
unwieldy. The yaml file is leaner and I think easier to peruse and edit.
* I obtained the 128 tests by scraping the output from running the
current doctests and munging with sed, grep, etc.. Going forward maybe
we can do something a little more robust - maybe from time-to-time run a
python script tbd to extract them from the docstrings in the same way
the existing doctests are extracted, compare to the existing list, and
add new ones.