Revise (slightly) Complex number implementation
Save precision, which is computed part of Complex numbers. Distinguish
exact value (the 2 parts of a complex number plus a erived "precision" value) from the often-used
and possibly approximate "value"; "value" is general the more useful and more
wanted quantity. In most computations, "value" is as good as, or
indistinguishable from the "exact_value".

Go over and make more correct some types.