Revise (slightly) Complex number implementation (#1552)
Save precision, which is a computed part of Complex numbers. Distinguish
exact value (the 2 parts of a complex number plus a derived "precision"
value) from the often-used and possibly approximate "value"; "value" is
generally the more useful and more wanted quantity. In most
computations, "value" is as good as, or indistinguishable from, the
"exact_value".

Go over and correct some types.